pyproject.toml
[project]
name = "paddle-pyfi"
version = "0.1.0"
description = "Reproducible PaddleOCR-VL + ERNIE 4.5 document intelligence pipeline with PyFi finance adapter"
readme = "README.md"
authors = [
    { name = "PrayerQX", email = "PrayerQX@users.noreply.github.com" }
]
requires-python = ">=3.12"
dependencies = [
    "openai>=2.32.0",
    "python-dotenv>=1.2.2",
    "pyyaml>=6.0.3",
    "requests>=2.33.1",
]

[project.scripts]
paddle-pyfi = "paddle_pyfi.cli:main"

[build-system]
requires = ["uv_build>=0.11.7,<0.12.0"]
build-backend = "uv_build"

[dependency-groups]
dev = [
    "pytest>=9.0.3",
]
